Clean up merge from dev to master
