+++ src/main/java/Sysc4806Group/demo/LoadDatabase.java
@@ -1,6 +1,5 @@
 package Sysc4806Group.demo;
 
-import Sysc4806Group.demo.entities.BookStore;
 import Sysc4806Group.demo.repositories.BookStoreRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -12,16 +11,15 @@ import org.springframework.context.annotation.Configuration;
 public class LoadDatabase {
 
     private static final Logger log = LoggerFactory.getLogger(LoadDatabase.class);
-
-    @Bean
-    CommandLineRunner initDatabase(BookStoreRepository repository){
-
-        BookStore bookstore = new BookStore();
-
-        return args -> {
-          log.info("Preloading..." + repository.save(bookstore));
-        };
-
-    }
+//
+//    @Bean
+//    CommandLineRunner initDatabase(BookStoreRepository repository){
+//
+//        BookStore bookstore = new BookStore();
+//
+//        return args -> {
+//          log.info("Preloading..." + repository.save(bookstore));
+//        };
+//    }
 
 }
