--- src/main/java/Sysc4806Group/demo/controllers/BookStoreController.java
@@ -1,7 +1,0 @@
-package Sysc4806Group.demo.controllers;
-
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class BookStoreController {
-}
