+++ src/main/java/Sysc4806Group/demo/entities/Book.java
@@ -1,12 +1,77 @@
 package Sysc4806Group.demo.entities;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.Id;
 
 @Entity
 public class Book {
 
     @Id
-    @GeneratedValue
-    private Long id = null;
+    private String isbn;
+    private String title;
+    private String author;
+    private String publisher;
+    private String pictureUrl;
+    private int inventory;
 
+    public Book(String isbn, String title, String author, String publisher, String pictureUrl, int inventory) {
+        this.isbn = isbn;
+        this.title = title;
+        this.author = author;
+        this.publisher = publisher;
+        this.pictureUrl = pictureUrl;
+        this.inventory = inventory;
+    }
+
+    public String getIsbn() {
+        return isbn;
+    }
+
+    public void setIsbn(String isbn) {
+        this.isbn = isbn;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getAuthor() {
+        return author;
+    }
+
+    public void setAuthor(String author) {
+        this.author = author;
+    }
+
+    public String getPublisher() {
+        return publisher;
+    }
+
+    public void setPublisher(String publisher) {
+        this.publisher = publisher;
+    }
+
+    public String getPictureUrl() {
+        return pictureUrl;
+    }
+
+    public void setPictureUrl(String pictureUrl) {
+        this.pictureUrl = pictureUrl;
+    }
+
+    public int getInventory() {
+        return inventory;
+    }
+
+    public void setInventory(int inventory) {
+        this.inventory = inventory;
+    }
+
+    public void sold() {
+        this.inventory -= 1;
+    }
 }
