+++ src/main/java/Sysc4806Group/demo/repositories/BookStoreRepository.java
@@ -1,7 +1,7 @@
 package Sysc4806Group.demo.repositories;
 
-import Sysc4806Group.demo.entities.BookStore;
+import Sysc4806Group.demo.entities.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookStoreRepository extends JpaRepository<BookStore, Long> {
+public interface BookStoreRepository extends JpaRepository<Book, Long> {
 }
